Move aggregation pushdown classes to aggregation package
This is to separate them from regular expression pushdown that will go
into `expression` package. Aggregation pushdown is related to general
expression pushdown and will build on top of it, but separating them a
bit will help navigate the codebase.
