+++ lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/aggregation/AggregateFunctionRewriter.java
@@ -11,11 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.base.expression;
+package io.trino.plugin.base.aggregation;
 
 import com.google.common.collect.ImmutableSet;
 import io.trino.matching.Match;
-import io.trino.plugin.base.expression.AggregateFunctionRule.RewriteContext;
+import io.trino.plugin.base.aggregation.AggregateFunctionRule.RewriteContext;
 import io.trino.spi.connector.AggregateFunction;
 import io.trino.spi.connector.ColumnHandle;
 import io.trino.spi.connector.ConnectorSession;
