+++ plugin/trino-base-jdbc/src/main/java/io/trino/plugin/jdbc/aggregation/ImplementCovarianceSamp.java
@@ -11,12 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.jdbc.expression;
+package io.trino.plugin.jdbc.aggregation;
 
 import io.trino.matching.Capture;
 import io.trino.matching.Captures;
 import io.trino.matching.Pattern;
-import io.trino.plugin.base.expression.AggregateFunctionRule;
+import io.trino.plugin.base.aggregation.AggregateFunctionRule;
 import io.trino.plugin.jdbc.JdbcColumnHandle;
 import io.trino.plugin.jdbc.JdbcExpression;
 import io.trino.spi.connector.AggregateFunction;
@@ -27,16 +27,16 @@ import java.util.Optional;
 
 import static com.google.common.base.Verify.verify;
 import static io.trino.matching.Capture.newCapture;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.basicAggregation;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.expressionTypes;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.functionName;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.inputs;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.variables;
+import static io.trino.plugin.base.aggregation.AggregateFunctionPatterns.basicAggregation;
+import static io.trino.plugin.base.aggregation.AggregateFunctionPatterns.expressionTypes;
+import static io.trino.plugin.base.aggregation.AggregateFunctionPatterns.functionName;
+import static io.trino.plugin.base.aggregation.AggregateFunctionPatterns.inputs;
+import static io.trino.plugin.base.aggregation.AggregateFunctionPatterns.variables;
 import static io.trino.spi.type.DoubleType.DOUBLE;
 import static io.trino.spi.type.RealType.REAL;
 import static java.lang.String.format;
 
-public class ImplementCovariancePop
+public class ImplementCovarianceSamp
         implements AggregateFunctionRule<JdbcExpression>
 {
     private static final Capture<List<Variable>> INPUTS = newCapture();
@@ -45,7 +45,7 @@ public class ImplementCovariancePop
     public Pattern<AggregateFunction> getPattern()
     {
         return basicAggregation()
-                .with(functionName().equalTo("covar_pop"))
+                .with(functionName().equalTo("covar_samp"))
                 .with(inputs().matching(
                         variables()
                                 .matching(expressionTypes(REAL, REAL).or(expressionTypes(DOUBLE, DOUBLE)))
@@ -63,7 +63,7 @@ public class ImplementCovariancePop
         verify(aggregateFunction.getOutputType().equals(columnHandle1.getColumnType()));
 
         return Optional.of(new JdbcExpression(
-                format("covar_pop(%s, %s)", context.getIdentifierQuote().apply(columnHandle1.getColumnName()), context.getIdentifierQuote().apply(columnHandle2.getColumnName())),
+                format("covar_samp(%s, %s)", context.getIdentifierQuote().apply(columnHandle1.getColumnName()), context.getIdentifierQuote().apply(columnHandle2.getColumnName())),
                 columnHandle1.getJdbcTypeHandle()));
     }
 }
