+++ plugin/trino-pinot/src/main/java/io/trino/plugin/pinot/query/aggregation/ImplementCountAll.java
@@ -15,17 +15,17 @@ package io.trino.plugin.pinot.query.aggregation;
 
 import io.trino.matching.Captures;
 import io.trino.matching.Pattern;
-import io.trino.plugin.base.expression.AggregateFunctionRule;
+import io.trino.plugin.base.aggregation.AggregateFunctionRule;
 import io.trino.plugin.pinot.query.AggregateExpression;
 import io.trino.spi.connector.AggregateFunction;
 
 import java.util.List;
 import java.util.Optional;
 
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.basicAggregation;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.functionName;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.inputs;
-import static io.trino.plugin.base.expression.AggregateFunctionPatterns.outputType;
+import static io.trino.plugin.base.aggregation.AggregateFunctionPatterns.basicAggregation;
+import static io.trino.plugin.base.aggregation.AggregateFunctionPatterns.functionName;
+import static io.trino.plugin.base.aggregation.AggregateFunctionPatterns.inputs;
+import static io.trino.plugin.base.aggregation.AggregateFunctionPatterns.outputType;
 import static io.trino.spi.type.BigintType.BIGINT;
 
 /**
