+++ plugin/trino-postgresql/src/main/java/io/trino/plugin/postgresql/ImplementAvgBigint.java
@@ -13,7 +13,7 @@
  */
 package io.trino.plugin.postgresql;
 
-import io.trino.plugin.jdbc.expression.BaseImplementAvgBigint;
+import io.trino.plugin.jdbc.aggregation.BaseImplementAvgBigint;
 
 public class ImplementAvgBigint
         extends BaseImplementAvgBigint
