+++ plugin/trino-sqlserver/src/main/java/io/trino/plugin/sqlserver/ImplementAvgBigint.java
@@ -13,7 +13,7 @@
  */
 package io.trino.plugin.sqlserver;
 
-import io.trino.plugin.jdbc.expression.BaseImplementAvgBigint;
+import io.trino.plugin.jdbc.aggregation.BaseImplementAvgBigint;
 
 public class ImplementAvgBigint
         extends BaseImplementAvgBigint
